Added new sentiment options.
After reviewing evaluation results, I reworked the sentiment analysis to also be able to look at an email as a whole instead of per sentence. This improved the results dramatically. Additionally, I found some decent options for where to draw the line for negative, neutral, and positive. These are values are subject to change pending further evaluation testing.