Finished calculation for time between emails.
This calculation currently returns a 2-d array where the values are stored by sent / received and then by day. User's email is currently calculated by highest frequency  of occurrence. Hour values are doubles with the whole number portion equal to the hours, and the decimal portion is equal to the minutes converted into a decimal value out of one hour.