Test for heatmap completed.
This set of tests checks to make sure the emails show up in the proper slot. It does so by sweeping on the day's day column and slotting emails into each hour. Then it does the same process for each day slot. Then one more test for the current time to make sure it matches.